minor bug fix, functions update

+++ lib/screens/post_list/post.view_header.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 import 'package:flutterpress/screens/post_list/mine_button.dart';
 import 'package:flutterpress/services/app.service.dart';
 import 'package:flutterpress/defines.dart';
@@ -22,26 +23,27 @@ class PostViewHeader extends StatelessWidget {
   Widget build(BuildContext context) {
     return Column(
       children: [
-        Container(
-          margin: EdgeInsets.all(sm),
-          child: Center(
-            child: ConstrainedBox(
-              constraints: BoxConstraints(
-                maxHeight: (titleSize * 3) + sm,
-                maxWidth: double.infinity,
-              ),
-              child: Text(
-                post.title,
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  color: Colors.black,
-                  fontSize: 20,
-                  fontWeight: FontWeight.w400,
+        if (!post.deleted)
+          Container(
+            margin: EdgeInsets.all(sm),
+            child: Center(
+              child: ConstrainedBox(
+                constraints: BoxConstraints(
+                  maxHeight: (titleSize * 3) + sm,
+                  maxWidth: double.infinity,
+                ),
+                child: Text(
+                  post.title,
+                  textAlign: TextAlign.center,
+                  style: TextStyle(
+                    color: Colors.black,
+                    fontSize: 20,
+                    fontWeight: FontWeight.w400,
+                  ),
                 ),
               ),
             ),
           ),
-        ),
         Container(
           margin: EdgeInsets.only(bottom: xs, left: md, right: md, top: xs),
           child: Row(
@@ -52,36 +54,47 @@ class PostViewHeader extends StatelessWidget {
                 width: 40,
               ),
               SizedBox(width: xs),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      '${AppService.authorName(post.authorName)} - No. ${post.id} - Date: ${post.date}',
-                      style: TextStyle(fontSize: sm),
-                    ),
+              if (!post.deleted) ...[
+                Expanded(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(
+                        '${AppService.authorName(post.authorName)} - No. ${post.id} - Date: ${post.date}',
+                        style: TextStyle(fontSize: sm),
+                      ),
 
-                    /// TODO: view count
-                    Text(
-                      post.slug + ' - 201 views',
-                      style: TextStyle(fontSize: md),
-                    )
-                  ],
+                      /// TODO: view count
+                      Text(
+                        post.slug + ' - 201 views',
+                        style: TextStyle(fontSize: md),
+                      )
+                    ],
+                  ),
                 ),
-              ),
-              if (AppService.isMine(post))
-                MineButton(
-                  onSelect: (option) {
-                    if (option == MineOption.update &&
-                        onUpdateButtonTap != null) {
-                      onUpdateButtonTap();
-                    }
-                    if (option == MineOption.delete &&
-                        onDeleteButtonTap != null) {
-                      onDeleteButtonTap();
-                    }
-                  },
-                ),
+                if (AppService.isMine(post))
+                  MineButton(
+                    onSelect: (option) {
+                      if (option == MineOption.update &&
+                          onUpdateButtonTap != null) {
+                        onUpdateButtonTap();
+                      }
+                      if (option == MineOption.delete &&
+                          onDeleteButtonTap != null) {
+                        onDeleteButtonTap();
+                      }
+                    },
+                  ),
+              ],
+              if (post.deleted)
+                Text(
+                  'deleted'.tr,
+                  style: TextStyle(
+                    fontSize: 20,
+                    color: Color(0x99000000),
+                    fontWeight: FontWeight.w500,
+                  ),
+                )
             ],
           ),
         )
